refactor(#339): 타입 선언 파일 변경
